refactor: break out request helpers into new classes from app.py
