+++ src/aws/osml/model_runner/common/credentials_utils.py
@@ -4,7 +4,7 @@ from typing import Dict
 
 import boto3
 
-from aws.osml.model_runner.app_config import BotoConfig
+from aws.osml.model_runner.config import BotoConfig
 
 from .exceptions import InvalidAssumedRoleException
