+++ test/aws/osml/model_runner/database/test_endpoint_statistics_table.py
@@ -25,7 +25,7 @@ class TestEndpointStatisticsTable(unittest.TestCase):
         """
         Set up virtual DDB resources/tables for each test to use
         """
-        from aws.osml.model_runner.app_config import BotoConfig
+        from aws.osml.model_runner.config import BotoConfig
         from aws.osml.model_runner.database.endpoint_statistics_table import EndpointStatisticsItem, EndpointStatisticsTable
 
         # Prepare something ahead of all tests
