+++ test/aws/osml/model_runner/database/test_job_table.py
@@ -24,7 +24,7 @@ class TestJobTable(unittest.TestCase):
         """
         Set up virtual DDB resources/tables for each test to use
         """
-        from aws.osml.model_runner.app_config import BotoConfig
+        from aws.osml.model_runner.config import BotoConfig
         from aws.osml.model_runner.database.job_table import JobItem, JobTable
 
         # Prepare something ahead of all tests
